Fix integer underflow when writing updated sk
This caused secret key files to become close to MAXINT bytes, as
the unsigned int that is the private key size would be subtracted
from MAXINT when its negative was used as offset.

+++ test/xmss_sign.c
@@ -43,7 +43,7 @@ int main(int argc, char **argv) {
     fread(m, 1, MLEN, stdin);
     xmss_core_sign(&params, sk, sm, &smlen, m, MLEN);
 
-    fseek(keypair, -params.privatekey_bytes, SEEK_CUR);
+    fseek(keypair, -((long int)params.privatekey_bytes), SEEK_CUR);
     fwrite(sk, 1, params.privatekey_bytes, keypair);
     fwrite(sm, 1, params.bytes + MLEN, stdout);
